+++ src/Exception/IllegalArgumentException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace NumPHP\Exception;
+
+class IllegalArgumentException extends \Exception
+{
+}