+++ src/Exception/MissingArgumentException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace NumPHP\Exception;
+
+class MissingArgumentException extends \Exception
+{
+}