+++ src/NumArray/FactoryInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace NumPHP\NumArray;
+
+interface FactoryInterface
+{
+    public function createFromData(array $data): NumArray;
+
+    public function createZeros(int ...$axis): NumArray;
+}