+++ CHANGELOG.md
@@ -4,15 +4,21 @@ All Notable changes to `NumPHP` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-## NEXT - 2017-01-27
+## NEXT - 2017-01-31
 
 ### Added
+- Add class `NumPHP\Exception\IllegalArgumentException`
+- Add class `NumPHP\Exception\MissingArgumentException`
 - Add class `NumPHP\NumArray\String\DefaultFormatter`
  - Add function `numArrayToString`
 - Add interface `NumPHP\NumArray\String\FormatterInterface`
  - Add function `numArrayToString`
 - Add class `NumPHP\NumArray\Factory`
  - Add function `createFromData`
+ - Add function `createZeros`
+- Add interface `NumPHP\NumArray\FactoryInterface`
+ - Add function `createFromData`
+ - Add function `createZeros`
 - Add class `NumPHP\NumArray\NumArray`
  - Add function `__toString`
  - Add function `getData`